Small refactor of EditorNavMenu into folder
