+++ editor.planx.uk/src/components/EditorNavMenu/EditorNavMenu.tsx
@@ -11,92 +11,14 @@ import RateReviewIcon from "@mui/icons-material/RateReview";
 import SchoolIcon from "@mui/icons-material/School";
 import TuneIcon from "@mui/icons-material/Tune";
 import Box from "@mui/material/Box";
-import IconButton from "@mui/material/IconButton";
-import { styled } from "@mui/material/styles";
 import Tooltip from "@mui/material/Tooltip";
-import Typography from "@mui/material/Typography";
-import { Role } from "@opensystemslab/planx-core/types";
 import { useStore } from "pages/FlowEditor/lib/store";
 import React, { useRef } from "react";
 import { useCurrentRoute, useLoadingRoute, useNavigation } from "react-navi";
-import { FONT_WEIGHT_SEMI_BOLD } from "theme";
 import EditorIcon from "ui/icons/Editor";
 
-interface Route {
-  title: string;
-  Icon: React.ElementType;
-  route: string;
-  accessibleBy: Role[];
-  disabled?: boolean;
-}
-
-interface RoutesForURL {
-  routes: Route[];
-  compact: boolean;
-}
-
-const MENU_WIDTH_COMPACT = "51px";
-const MENU_WIDTH_FULL = "164px";
-
-const Root = styled(Box, {
-  shouldForwardProp: (prop) => prop !== "compact",
-})<{ compact?: boolean }>(({ theme, compact }) => ({
-  width: compact ? MENU_WIDTH_COMPACT : MENU_WIDTH_FULL,
-  flexShrink: 0,
-  background: theme.palette.background.paper,
-  borderRight: `1px solid ${theme.palette.border.light}`,
-}));
-
-const MenuWrap = styled("ul")(({ theme }) => ({
-  listStyle: "none",
-  margin: 0,
-  padding: theme.spacing(1, 0.4, 0, 0.4),
-  position: "sticky",
-  top: 0,
-}));
-
-const MenuItem = styled("li")(({ theme }) => ({
-  margin: theme.spacing(0.75, 0),
-  padding: 0,
-}));
-
-const MenuTitle = styled(Typography)(({ theme }) => ({
-  fontWeight: FONT_WEIGHT_SEMI_BOLD,
-  paddingLeft: theme.spacing(0.5),
-  textAlign: "left",
-})) as typeof Typography;
-
-const MenuButton = styled(IconButton, {
-  shouldForwardProp: (prop) => prop !== "isActive",
-})<{ isActive: boolean }>(({ theme, isActive, disabled }) => ({
-  color: theme.palette.text.primary,
-  width: "100%",
-  border: "1px solid transparent",
-  justifyContent: "flex-start",
-  borderRadius: "3px",
-  "&:hover": {
-    background: theme.palette.common.white,
-    borderColor: theme.palette.border.light,
-  },
-  ...(isActive && {
-    background: theme.palette.common.white,
-    color: theme.palette.text.primary,
-    border: `1px solid ${theme.palette.border.main}`,
-    "&:hover": {
-      borderColor: theme.palette.border.main,
-    },
-  }),
-  ...(disabled && {
-    color: theme.palette.text.disabled,
-    "&:hover": {
-      background: "none",
-      borderColor: "transparent",
-    },
-  }),
-  "& > svg": {
-    opacity: 0.8,
-  },
-}));
+import { MenuButton, MenuItem, MenuTitle, MenuWrap, Root } from "./styles";
+import { Route, RoutesForURL } from "./types";
 
 function EditorNavMenu() {
   const { navigate } = useNavigation();
