+++ editor.planx.uk/src/components/EditorNavMenu/styles.tsx
@@ -1,0 +1,86 @@
+import AdminPanelSettingsIcon from "@mui/icons-material/AdminPanelSettings";
+import AssignmentTurnedInIcon from "@mui/icons-material/AssignmentTurnedIn";
+import FactCheckIcon from "@mui/icons-material/FactCheck";
+import FormatListBulletedIcon from "@mui/icons-material/FormatListBulleted";
+import GroupIcon from "@mui/icons-material/Group";
+import Info from "@mui/icons-material/Info";
+import LeaderboardIcon from "@mui/icons-material/Leaderboard";
+import MenuBookIcon from "@mui/icons-material/MenuBook";
+import PaletteIcon from "@mui/icons-material/Palette";
+import RateReviewIcon from "@mui/icons-material/RateReview";
+import SchoolIcon from "@mui/icons-material/School";
+import TuneIcon from "@mui/icons-material/Tune";
+import Box from "@mui/material/Box";
+import IconButton from "@mui/material/IconButton";
+import { styled } from "@mui/material/styles";
+import Tooltip from "@mui/material/Tooltip";
+import Typography from "@mui/material/Typography";
+import { Role } from "@opensystemslab/planx-core/types";
+import { useStore } from "pages/FlowEditor/lib/store";
+import React, { useRef } from "react";
+import { useCurrentRoute, useLoadingRoute, useNavigation } from "react-navi";
+import { FONT_WEIGHT_SEMI_BOLD } from "theme";
+import EditorIcon from "ui/icons/Editor";
+
+const MENU_WIDTH_COMPACT = "51px";
+const MENU_WIDTH_FULL = "164px";
+
+export const Root = styled(Box, {
+  shouldForwardProp: (prop) => prop !== "compact",
+})<{ compact?: boolean }>(({ theme, compact }) => ({
+  width: compact ? MENU_WIDTH_COMPACT : MENU_WIDTH_FULL,
+  flexShrink: 0,
+  background: theme.palette.background.paper,
+  borderRight: `1px solid ${theme.palette.border.light}`,
+}));
+
+export const MenuWrap = styled("ul")(({ theme }) => ({
+  listStyle: "none",
+  margin: 0,
+  padding: theme.spacing(1, 0.4, 0, 0.4),
+  position: "sticky",
+  top: 0,
+}));
+
+export const MenuItem = styled("li")(({ theme }) => ({
+  margin: theme.spacing(0.75, 0),
+  padding: 0,
+}));
+
+export const MenuTitle = styled(Typography)(({ theme }) => ({
+  fontWeight: FONT_WEIGHT_SEMI_BOLD,
+  paddingLeft: theme.spacing(0.5),
+  textAlign: "left",
+})) as typeof Typography;
+
+export const MenuButton = styled(IconButton, {
+  shouldForwardProp: (prop) => prop !== "isActive",
+})<{ isActive: boolean }>(({ theme, isActive, disabled }) => ({
+  color: theme.palette.text.primary,
+  width: "100%",
+  border: "1px solid transparent",
+  justifyContent: "flex-start",
+  borderRadius: "3px",
+  "&:hover": {
+    background: theme.palette.common.white,
+    borderColor: theme.palette.border.light,
+  },
+  ...(isActive && {
+    background: theme.palette.common.white,
+    color: theme.palette.text.primary,
+    border: `1px solid ${theme.palette.border.main}`,
+    "&:hover": {
+      borderColor: theme.palette.border.main,
+    },
+  }),
+  ...(disabled && {
+    color: theme.palette.text.disabled,
+    "&:hover": {
+      background: "none",
+      borderColor: "transparent",
+    },
+  }),
+  "& > svg": {
+    opacity: 0.8,
+  },
+}));
